flask-ocr-service/requirements.txt
flask
flask-cors
pytesseract
pillow
opencv-python-headless
numpy
